for clicks within the viewport, now a picker cone gets determined

+++ src/pointcloud_viewer/navigation.hpp
@@ -49,7 +49,7 @@ protected:
 signals:
   void mouse_sensitivity_value_changed(int value);
 
-  void simpleLeftClick(glm::vec2 point);
+  void simpleLeftClick(glm::ivec2 point);
 
 private:
   enum mode_t
